+++ crabzilla-example1/crabzilla-example1-vertx/src/main/java/crabzilla/example1/Example1VertxLauncher.java
@@ -6,8 +6,8 @@ import crabzilla.example1.aggregates.customer.Customer;
 import crabzilla.example1.aggregates.customer.CustomerId;
 import crabzilla.example1.aggregates.customer.commands.CreateCustomerCmd;
 import crabzilla.model.UnitOfWork;
-import crabzilla.stacks.vertx.codecs.gson.CommandCodec;
-import crabzilla.stacks.vertx.verticles.CommandHandlerVerticle;
+import crabzilla.stack.vertx.codecs.gson.CommandCodec;
+import crabzilla.stack.vertx.verticles.CommandHandlerVerticle;
 import io.vertx.core.Vertx;
 import io.vertx.core.VertxOptions;
 import io.vertx.core.eventbus.DeliveryOptions;
