+++ crabzilla-example1/crabzilla-example1-vertx/src/main/java/crabzilla/example1/Example1VertxModule.java
@@ -26,13 +26,13 @@ import crabzilla.example1.services.SampleServiceImpl;
 import crabzilla.model.*;
 import crabzilla.stack.EventRepository;
 import crabzilla.stack.SnapshotReaderFn;
-import crabzilla.stacks.vertx.codecs.fst.AggregateRootIdCodec;
-import crabzilla.stacks.vertx.codecs.fst.CommandCodec;
-import crabzilla.stacks.vertx.codecs.fst.EventCodec;
-import crabzilla.stacks.vertx.codecs.fst.UnitOfWorkCodec;
-import crabzilla.stacks.vertx.gson.RuntimeTypeAdapterFactory;
-import crabzilla.stacks.vertx.sql.JdbiEventRepository;
-import crabzilla.stacks.vertx.verticles.CommandHandlerVerticle;
+import crabzilla.stack.vertx.codecs.fst.AggregateRootIdCodec;
+import crabzilla.stack.vertx.codecs.fst.CommandCodec;
+import crabzilla.stack.vertx.codecs.fst.EventCodec;
+import crabzilla.stack.vertx.codecs.fst.UnitOfWorkCodec;
+import crabzilla.stack.vertx.gson.RuntimeTypeAdapterFactory;
+import crabzilla.stack.vertx.sql.JdbiEventRepository;
+import crabzilla.stack.vertx.verticles.CommandHandlerVerticle;
 import io.vertx.core.Vertx;
 import lombok.val;
 import net.dongliu.gson.GsonJava8TypeAdapterFactory;
