+++ crabzilla-vertx-stack/src/main/java/crabzilla/stack/vertx/codecs/fst/UnitOfWorkCodec.java
@@ -1,25 +1,25 @@
-package crabzilla.stacks.vertx.codecs.fst;
+package crabzilla.stack.vertx.codecs.fst;
 
-import crabzilla.model.Command;
+import crabzilla.model.UnitOfWork;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.eventbus.MessageCodec;
 import org.nustaq.serialization.FSTConfiguration;
 
 import javax.inject.Inject;
 
-public class CommandCodec implements MessageCodec<Command, Command> {
+public class UnitOfWorkCodec implements MessageCodec<UnitOfWork, UnitOfWork> {
 
   final FSTConfiguration fst;
 
   @Inject
-  public CommandCodec(FSTConfiguration fst) {
+  public UnitOfWorkCodec(FSTConfiguration fst) {
     this.fst = fst;
   }
 
   @Override
-  public void encodeToWire(Buffer buffer, Command Command) {
+  public void encodeToWire(Buffer buffer, UnitOfWork UnitOfWork) {
 
-    final byte barray[] = fst.asByteArray(Command);
+    final byte barray[] = fst.asByteArray(UnitOfWork);
 
     // Write data into given buffer
     buffer.appendInt(barray.length);
@@ -27,7 +27,7 @@ public class CommandCodec implements MessageCodec<Command, Command> {
   }
 
   @Override
-  public Command decodeFromWire(int pos, Buffer buffer) {
+  public UnitOfWork decodeFromWire(int pos, Buffer buffer) {
 
     // My custom message starting from this *position* of buffer
     int _pos = pos;
@@ -46,12 +46,12 @@ public class CommandCodec implements MessageCodec<Command, Command> {
       throw new RuntimeException("When decodingFromWire", e);
     }
 
-    return (Command) readObj;
+    return (UnitOfWork) readObj;
   }
 
   @Override
-  public Command transform(Command Command) {
-    return Command;
+  public UnitOfWork transform(UnitOfWork UnitOfWork) {
+    return UnitOfWork;
   }
 
   @Override
