+++ crabzilla-vertx-stack/src/main/java/crabzilla/stack/vertx/verticles/EventsProjectionVerticle.java
@@ -1,4 +1,4 @@
-package crabzilla.stacks.vertx.verticles;
+package crabzilla.stack.vertx.verticles;
 
 public class EventsProjectionVerticle {
 }
